+++ internal/server/contentmigrate_test.go
@@ -129,6 +129,7 @@ var notIDs = map[string]string{
 	"server.containerFile.Paintings[].Y":                whyPos,
 	"server.containerFile.Paintings[].Z":                whyPos,
 	"server.containerFile.ShelfLast{}":                  whyCount,
+	"server.containerFile.SpawnerDelays{}":              whyCount,
 	"server.containerFile.Stands[].Dim":                 whyPos,
 	"server.containerFile.Stands[].Fire":                whyCount,
 	"server.containerFile.Stars{}[].Colors[]":           whyRgb,